GUI-1281 Update derived column extension

+++ assets/extensions/ext-unused/derivedConversion.ext.js
@@ -4,7 +4,12 @@ window.UExtDerivedConversion = (function(UExtDerivedConversion) {
     UExtDerivedConversion.buttons = [{
         "buttonText": "Derived Field Conversion",
         "fnName": "derivedConversion",
-        "arrayOfFields": [],
+        "arrayOfFields": [{
+            "name": "Keep original",
+            "fieldClass": "keep",
+            "type": "boolean",
+            "autofill": true
+        }],
     }];
 
     UExtDerivedConversion.actionFn = function(functionName) {
@@ -16,51 +21,6 @@ window.UExtDerivedConversion = (function(UExtDerivedConversion) {
                 return null;
         }
     };
-
-    function getDerivedCol(ext, tableInfo, col, srcTableName) {
-        var deferred = XcSDK.Promise.deferred();
-
-        var tableName = tableInfo.name;
-
-        if (col.type === 'array' || col.type === 'object') {
-            // array and object columns will be projected away at the end
-            // this case also handles 'DATA' column, and leaves table unchanged
-            deferred.resolve(tableName);
-        } else if (!col.prefix) {
-            // derived fields do not need to be processed, but must be added to
-            // the columns projected at the end
-            tableInfo.colsToProject.push(col.backName);
-            deferred.resolve(tableName);
-        } else {
-            // convert prefix field of primitive type to derived
-            var mapFn;
-            if (col.type === 'integer' || col.type === 'float') {
-                // convert all numbers to float, since we cannot determine
-                // actual type of prefix fields
-                mapFn = "float";
-            } else if (col.type === 'boolean') {
-                mapFn = "bool";
-            } else {
-                mapFn = "string";
-            }
-            var mapStr = mapFn + "(" + col.backName + ")";
-            var newColName = ext.createUniqueCol(srcTableName,
-                col.backName.replace('::', '_').replace(/\./g, '_'), false);
-
-            ext.map(mapStr, tableName, newColName, ext.createTempTableName())
-            .then(function(tableAfterMap) {
-                // add new column to array of columns to be projected finally
-                tableInfo.colsToProject.push(newColName);
-                // update latest table name so that all modified columns end
-                // up in a single table
-                tableInfo.name = tableAfterMap;
-                deferred.resolve(tableAfterMap);
-            })
-            .fail(deferred.reject);
-        }
-
-        return deferred.promise();
-    }
 
     /*
     convertToDerived: Converts prefix fields to derived fields and removes object
@@ -76,33 +36,79 @@ window.UExtDerivedConversion = (function(UExtDerivedConversion) {
             var self = this;
             var srcTable = self.getTriggerTable();
             var srcTableName = srcTable.getName();
+            var args = self.getArgs();
+            var keep = args.keep;
 
             var cols = srcTable.tableCols;
-            var promises = [];
-            var tableInfo = {"name": srcTableName, "colsToProject": []};
+            var mapStrs = [];
+            var newColNames = [];
+            var colsToProject = [];
+            var usedCols = {};
             // each promise processes one col of the table
             for (var i = 0; i < cols.length; i++) {
                 var col = cols[i];
-                promises.push(getDerivedCol.bind(window, ext, tableInfo,
-                    col, srcTableName));
+                if (col.type === 'array' || col.type === 'object') {
+                    // array and object columns will be projected away at the end
+                    // this case also handles 'DATA' column, and leaves table unchanged
+                    continue;
+                } else if (!col.prefix) {
+                    // derived fields do not need to be processed, but must be added to
+                    // the columns projected at the end
+                    colsToProject.push(col.backName);
+                    usedCols[col.backName] = true;
+                    continue;
+                } else {
+                    // convert prefix field of primitive type to derived
+                    var mapFn;
+                    if (col.type === 'integer' || col.type === 'float') {
+                        // convert all numbers to float, since we cannot determine
+                        // actual type of prefix fields
+                        mapFn = "float";
+                    } else if (col.type === 'boolean') {
+                        mapFn = "bool";
+                    } else {
+                        mapFn = "string";
+                    }
+                    var mapStr = mapFn + "(" + col.backName + ")";
+                    var newColName = ext.convertPrefixColumn(col.backName);
+                    if (usedCols.hasOwnProperty(newColName)) {
+                        // remove duplicate name
+                        newColName = ext.createUniqueCol(srcTableName, newColName, false);
+                    }
+
+                    mapStrs.push(mapStr);
+                    newColNames.push(newColName);
+                    usedCols[newColName] = true;
+                    colsToProject.push(newColName);
+                }
             }
 
-            XcSDK.Promise.chain(promises)
+            var def = mapStrs.length === 0 ?
+            XcSDK.Promise.resolve(srcTableName) :
+            ext.map(mapStrs, srcTableName, newColNames);
+
+            def
             .then(function(derivedTable) {
-                // project the processed prefix columns and the original
-                // original derived columns
-                var newTableName = ext.createTableName(null, null, srcTableName);
-                return ext.project(tableInfo.colsToProject, derivedTable,
-                    newTableName);
+                if (keep) {
+                    return XcSDK.Promise.resolve(derivedTable);
+                } else {
+                    // project the processed prefix columns and the original
+                    // original derived columns
+                    var newTableName = ext.createTableName(null, null, srcTableName);
+                    return ext.project(colsToProject, derivedTable, newTableName);
+                }
             })
             .then(function(projectedTable) {
+                if (srcTableName === projectedTable) {
+                    return;
+                }
                 // hide all columns and display only the one's projected
                 var table = ext.getTable(projectedTable);
                 table.deleteAllCols();
-                tableInfo.colsToProject.forEach(function(colName) {
+                colsToProject.forEach(function(colName) {
                     table.addCol(new XcSDK.Column(colName));
                 });
-                return table.addToWorksheet();
+                return table.addToWorksheet(srcTableName);
             })
             .then(deferred.resolve)
             .fail(deferred.reject);
